dam/requirements.txt
pytorch~=2.6.0
pysoundfile~=0.13.1
torchaudio~=2.7.0
transformers~=4.52.3
peft~=0.15.2
